test: use real httpflow for firewall metadata logs
